requirements.txt
bcrypt == 3.1.6
cffi == 1.12.3
Click == 7.0
Faker == 1.0.7
Flask == 1.0.3
Flask-Bcrypt == 0.7.1
Flask-Login == 0.4.1
Flask-WTF == 0.14.2
itsdangerous == 1.1.0
Jinja2 == 2.10.1
MarkupSafe == 1.1.1
peewee == 3.9.5
pycparser == 2.19
python-dateutil == 2.8.0
python-slugify == 3.0.2
six == 1.12.0
text-unidecode == 1.2
Werkzeug == 0.15.4
WTForms == 2.2.1
